Made several refactoring changes * Changed properties to ivars where possible for consistency * Avoid property self.x and .x= syntax where the method is custom implemented in the class * Changed outlets to be strong * Changed other assign properties to be weak * Avoid use of "int" and enums with non-specified type * Changed KVO access of registers view controller to use a delegate instead

+++ Bit Slicer/ZGDebugThread.h
@@ -36,9 +36,10 @@
 
 @interface ZGDebugThread : NSObject
 
-- (id)initWithThread:(thread_act_t)thread registerIndex:(int)registerIndex;
+// registerIndex should only be within [0, 4)
+- (id)initWithThread:(thread_act_t)thread registerIndex:(uint8_t)registerIndex;
 
 @property (readonly, nonatomic) thread_act_t thread;
-@property (readonly, nonatomic) int registerIndex;
+@property (readonly, nonatomic) uint8_t registerIndex;
 
 @end
